# [Core] Scheduling optimization 2