@rkooo567 2024-04-23
this commit: Throughput: 24.59 requests/s, 12587.93 tokens/s
before regression: Throughput: 24.92 requests/s, 12757.61 tokens/s
before this commit: Throughput: 24.23 requests/s, 12404.63 tokens/s

So less than 2% diff