@rkooo567 2024-04-23
```
Throughput: 24.80 requests/s, 12699.84 tokens/s
```

lol I added if opposite when I removed the flag and replaced it to enable_chunk

After fixing this, the perf seems almost identical to before regrssion @simon-mo